Creates balanced design by removing overlapping fractions and filling incomplete rows

Creates balanced design by removing overlapping fractions and filling incomplete rows

MSstatsBalancedDesign(
  input,
  feature_columns,
  fill_incomplete = TRUE,
  handle_fractions = TRUE,
  fix_missing = NULL
)

Arguments

input
data.table processed by the MSstatsPreprocess function

feature_columns
str, names of columns that define spectral features

fill_incomplete
if TRUE (default), Intensity values for missing runs will be added as NA

handle_fractions
if TRUE (default), overlapping fractions will be resolved

fix_missing
str, optional. Defaults to NULL, which means no action. If not NULL, must be one of the options: "zero_to_na" or "na_to_zero". If "zero_to_na", Intensity values equal exactly to 0 will be converted to NA. If "na_to_zero", missing values will be replaced by zeros.

Value

data.frame of class MSstatsValidated

Examples

unbalanced_data = system.file("tinytest/raw_data/unbalanced_data.csv",
                              package = "MSstatsConvert")
unbalanced_data = data.table::as.data.table(read.csv(unbalanced_data))
balanced = MSstatsBalancedDesign(unbalanced_data,
                                 c("PeptideSequence", "PrecursorCharge",
                                   "FragmentIon", "ProductCharge"))

#> INFO  [2021-05-10 23:03:38] ** Features with one or two measurements across runs are removed.
#> INFO  [2021-05-10 23:03:38] ** Fractionation handled.
#> INFO  [2021-05-10 23:03:38] ** Updated quantification data to make balanced design. Missing values are marked by NA

dim(balanced) # Now balanced has additional rows (with Intensity = NA)

#> [1] 721  11

# for runs that were not included in the unbalanced_data table