Import files from signal processing tools.

Import files from signal processing tools.

MSstatsImport(input_files, type, tool, tool_version = NULL, ...)

Arguments

input_files
list of paths to input files or data.frame objects. Interpretation of this parameter depends on values of parameters type and tool.

type
chr, "MSstats" or "MSstatsTMT".

tool
chr, name of a signal processing tool that generated input files.

tool_version
not implemented yet. In the future, this parameter will allow handling different versions of each signal processing tools.

...
optional additional parameters to data.table::fread.

Value

an object of class MSstatsInputFiles.

Examples

evidence_path = system.file("tinytest/raw_data/MaxQuant/mq_ev.csv",
                            package = "MSstatsConvert")
pg_path = system.file("tinytest/raw_data/MaxQuant/mq_pg.csv",
                      package = "MSstatsConvert")
evidence = read.csv(evidence_path)
pg = read.csv(pg_path)
imported = MSstatsImport(list(evidence = evidence, protein_groups = pg),
                         "MSstats", "MaxQuant")

#> INFO  [2021-05-05 12:54:41] ** Raw data from MaxQuant imported successfully.

class(imported)

#> [1] "MSstatsMaxQuantFiles"
#> attr(,"package")
#> [1] "MSstatsConvert"

head(getInputFile(imported, "evidence"))

#>       Sequence Length Modifications Modifiedsequence OxidationMProbabilities
#> 1: AEAPAAAPAAK     11    Unmodified    _AEAPAAAPAAK_                        
#> 2: AEAPAAAPAAK     11    Unmodified    _AEAPAAAPAAK_                        
#> 3: AEAPAAAPAAK     11    Unmodified    _AEAPAAAPAAK_                        
#> 4: AEAPAAAPAAK     11    Unmodified    _AEAPAAAPAAK_                        
#> 5: AEAPAAAPAAK     11    Unmodified    _AEAPAAAPAAK_                        
#> 6: AEAPAAAPAAK     11    Unmodified    _AEAPAAAPAAK_                        
#>    OxidationMScoreDiffs AcetylProteinNterm OxidationM Missedcleavages Proteins
#> 1:                                       0          0               0   P06959
#> 2:                                       0          0               0   P06959
#> 3:                                       0          0               0   P06959
#> 4:                                       0          0               0   P06959
#> 5:                                       0          0               0   P06959
#> 6:                                       0          0               0   P06959
#>    LeadingProteins LeadingRazorProtein GeneNames
#> 1:          P06959              P06959      aceF
#> 2:          P06959              P06959      aceF
#> 3:          P06959              P06959      aceF
#> 4:          P06959              P06959      aceF
#> 5:          P06959              P06959      aceF
#> 6:          P06959              P06959      aceF
#>                                                                                 ProteinNames
#> 1: Dihydrolipoyllysine-residue acetyltransferase component of pyruvate dehydrogenase complex
#> 2: Dihydrolipoyllysine-residue acetyltransferase component of pyruvate dehydrogenase complex
#> 3: Dihydrolipoyllysine-residue acetyltransferase component of pyruvate dehydrogenase complex
#> 4: Dihydrolipoyllysine-residue acetyltransferase component of pyruvate dehydrogenase complex
#> 5: Dihydrolipoyllysine-residue acetyltransferase component of pyruvate dehydrogenase complex
#> 6: Dihydrolipoyllysine-residue acetyltransferase component of pyruvate dehydrogenase complex
#>          Type                                     Rawfile
#> 1: MULTI-MSMS 121219_S_CCES_01_01_LysC_Try_1to10_Mixt_1_1
#> 2: MULTI-MSMS 121219_S_CCES_01_02_LysC_Try_1to10_Mixt_1_2
#> 3: MULTI-MSMS 121219_S_CCES_01_03_LysC_Try_1to10_Mixt_1_3
#> 4: MULTI-MSMS 121219_S_CCES_01_05_LysC_Try_1to10_Mixt_2_2
#> 5: MULTI-MSMS 121219_S_CCES_01_06_LysC_Try_1to10_Mixt_2_3
#> 6: MULTI-MSMS 121219_S_CCES_01_08_LysC_Try_1to10_Mixt_3_2
#>                    Experiment MSMSmz Charge      mz     Mass Resolution
#> 1: 01_LysC_Try_1to10_Mixt_1_1 484.26      2 484.264 966.5134   60409.30
#> 2: 02_LysC_Try_1to10_Mixt_1_2 484.26      2 484.264 966.5134   61157.75
#> 3: 03_LysC_Try_1to10_Mixt_1_3 484.26      2 484.264 966.5134   61423.33
#> 4: 05_LysC_Try_1to10_Mixt_2_2 484.26      2 484.264 966.5134   60286.66
#> 5: 06_LysC_Try_1to10_Mixt_2_3 484.27      2 484.264 966.5134   60426.07
#> 6: 08_LysC_Try_1to10_Mixt_3_2 484.26      2 484.264 966.5134   61518.70
#>    UncalibratedCalibratedmzppm UncalibratedCalibratedmzDa MassErrorppm
#> 1:                    0.112640                 5.4549e-05   -0.0032325
#> 2:                    0.070686                 3.4231e-05    0.0570860
#> 3:                    0.222730                 1.0786e-04   -0.3117600
#> 4:                    0.027139                 1.3143e-05   -0.0047083
#> 5:                    0.207260                 1.0037e-04   -0.2215600
#> 6:                   -0.061896                -2.9974e-05    0.0102090
#>    MassErrorDa UncalibratedMassErrorppm UncalibratedMassErrorDa Maxintensitymz0
#> 1: -1.5654e-06                 0.109410              5.2984e-05        484.2639
#> 2:  2.7645e-05                 0.127770              6.1876e-05        484.2639
#> 3: -1.5098e-04                -0.089031             -4.3115e-05        484.2637
#> 4: -2.2801e-06                 0.022431              1.0863e-05        484.2639
#> 5: -1.0729e-04                -0.014302             -6.9259e-06        484.2637
#> 6:  4.9436e-06                -0.051687             -2.5030e-05        484.2639
#>    Retentiontime Retentionlength Calibratedretentiontime
#> 1:        17.858         0.49251                  17.858
#> 2:        18.200         0.41806                  17.844
#> 3:        17.872         0.53369                  18.072
#> 4:        18.619         0.49627                  17.717
#> 5:        18.043         0.51981                  18.193
#> 6:        18.902         0.45821                  17.700
#>    Calibratedretentiontimestart Calibratedretentiontimefinish
#> 1:                       17.562                        18.054
#> 2:                       17.613                        18.031
#> 3:                       17.794                        18.327
#> 4:                       17.443                        17.940
#> 5:                       17.922                        18.442
#> 6:                       17.415                        17.873
#>    Retentiontimecalibration Matchtimedifference Matchmzdifference Matchqvalue
#> 1:                  0.00000                  NA                NA          NA
#> 2:                 -0.35622                  NA                NA          NA
#> 3:                  0.19934                  NA                NA          NA
#> 4:                 -0.90250                  NA                NA          NA
#> 5:                  0.15029                  NA                NA          NA
#> 6:                 -1.20210                  NA                NA          NA
#>    Matchscore Numberofdatapoints Numberofscans Numberofisotopicpeaks PIF
#> 1:         NA                 52            18                     4   0
#> 2:         NA                 44            13                     4   0
#> 3:         NA                 41            15                     4   0
#> 4:         NA                 32             9                     5   0
#> 5:         NA                 34            12                     4   0
#> 6:         NA                 39            13                     4   0
#>    Fractionoftotalspectrum Basepeakfraction        PEP MSMSCount MSMSScanNumber
#> 1:                       0                0 0.00203910         1           2786
#> 2:                       0                0 0.00064103         1           2994
#> 3:                       0                0 0.00013657         2           3072
#> 4:                       0                0 0.00031023         1           2560
#> 5:                       0                0 0.00043291         1           2504
#> 6:                       0                0 0.00040251         1           2597
#>      Score Deltascore Combinatorics Intensity Reverse Potentialcontaminant   id
#> 1:  76.332     72.245             1   4023100      NA                   NA 6955
#> 2:  83.081     78.741             1   5132500      NA                   NA 6956
#> 3: 104.430    101.250             1   2761600      NA                   NA 6957
#> 4:  94.465     85.666             1   4091800      NA                   NA 6958
#> 5:  88.596     84.180             1   4727000      NA                   NA 6959
#> 6:  90.050     79.395             1   2258400      NA                   NA 6960
#>    ProteingroupIDs PeptideID ModpeptideID   MSMSIDs BestMSMS AIFMSMSIDs
#> 1:              92       289          316      6071     6071         NA
#> 2:              92       289          316      6072     6072         NA
#> 3:              92       289          316 6073;6074     6073         NA
#> 4:              92       289          316      6075     6075         NA
#> 5:              92       289          316      6076     6076         NA
#> 6:              92       289          316      6077     6077         NA
#>    OxidationMsiteIDs
#> 1:                  
#> 2:                  
#> 3:                  
#> 4:                  
#> 5:                  
#> 6: