Create annotation

Create annotation

MSstatsMakeAnnotation(input, annotation, ...)

Arguments

input
data.table preprocessed by the MSstatsClean function

annotation
data.table

...
key-value pairs, where keys are names of columns of annotation

Value

data.table

Examples

evidence_path = system.file("tinytest/raw_data/MaxQuant/mq_ev.csv",
                            package = "MSstatsConvert")
pg_path = system.file("tinytest/raw_data/MaxQuant/mq_pg.csv",
                      package = "MSstatsConvert")
evidence = read.csv(evidence_path)
pg = read.csv(pg_path)
imported = MSstatsImport(list(evidence = evidence, protein_groups = pg),
                         "MSstats", "MaxQuant")

#> INFO  [2021-05-10 23:03:41] ** Raw data from MaxQuant imported successfully.

cleaned_data = MSstatsClean(imported, protein_id_col = "Proteins")

#> INFO  [2021-05-10 23:03:41] ** Rows with values of Potentialcontaminant equal to + are removed 
#> INFO  [2021-05-10 23:03:41] ** Rows with values of Reverse equal to + are removed 
#> INFO  [2021-05-10 23:03:41] ** Rows with values of Potentialcontaminant equal to + are removed 
#> INFO  [2021-05-10 23:03:41] ** Rows with values of Reverse equal to + are removed 
#> INFO  [2021-05-10 23:03:41] ** + Contaminant, + Reverse, + Potential.contaminant proteins are removed.
#> INFO  [2021-05-10 23:03:41] ** Raw data from MaxQuant cleaned successfully.

annot_path = system.file("tinytest/raw_data/MaxQuant/annotation.csv",
                         package = "MSstatsConvert")
mq_annot = MSstatsMakeAnnotation(cleaned_data, read.csv(annot_path),
                                 Run = "Rawfile")

#> INFO  [2021-05-10 23:03:41] ** Using provided annotation.
#> INFO  [2021-05-10 23:03:41] ** Run labels were standardized to remove symbols such as '.' or '%'.

head(mq_annot)

#>                                            Run Condition BioReplicate
#> 1: 121219_S_CCES_01_01_LysC_Try_1to10_Mixt_1_1         1            1
#> 2: 121219_S_CCES_01_02_LysC_Try_1to10_Mixt_1_2         1            1
#> 3: 121219_S_CCES_01_03_LysC_Try_1to10_Mixt_1_3         1            1
#> 4: 121219_S_CCES_01_04_LysC_Try_1to10_Mixt_2_1         2            2
#> 5: 121219_S_CCES_01_05_LysC_Try_1to10_Mixt_2_2         2            2
#> 6: 121219_S_CCES_01_06_LysC_Try_1to10_Mixt_2_3         2            2
#>    Experiment IsotopeLabelType
#> 1:        1_1                L
#> 2:        1_2                L
#> 3:        1_3                L
#> 4:        2_1                L
#> 5:        2_2                L
#> 6:        2_3                L